refactor(passport): move files into cookie/, refresh/, account/, adapters/ subdirectories
Restructure packages/passport/src/ by responsibility:
- cookie/: cookie-auth.ts, cookie-auth-module.ts, cookie-manager.ts + tests
- refresh/: refresh-token.ts, jwt-refresh-token-adapter.ts + tests
- account/: account-linking.ts + tests
- adapters/: passport-js.ts

Update all internal relative import paths to match new locations.
index.ts re-export paths updated; public API signature unchanged.

Resolves #855
