+++ packages/passport/src/adapters/passport-js.ts
@@ -2,9 +2,9 @@ import type { Token } from '@konekti/core';
 import type { GuardContext, Principal } from '@konekti/http';
 import type { Provider } from '@konekti/di';
 
-import { AuthenticationFailedError, AuthenticationRequiredError } from './errors.js';
-import { normalizePrincipalScopes } from './scope.js';
-import type { AuthHandledResult, AuthStrategy, AuthStrategyRegistration } from './types.js';
+import { AuthenticationFailedError, AuthenticationRequiredError } from '../errors.js';
+import { normalizePrincipalScopes } from '../scope.js';
+import type { AuthHandledResult, AuthStrategy, AuthStrategyRegistration } from '../types.js';
 
 interface PassportJsActionBindings {
   error?: (error: unknown) => void;
