+++ packages/passport/src/index.ts
@@ -1,14 +1,14 @@
-export * from './account-linking.js';
-export * from './cookie-auth.js';
-export * from './cookie-auth-module.js';
-export * from './cookie-manager.js';
+export * from './account/account-linking.js';
+export * from './cookie/cookie-auth.js';
+export * from './cookie/cookie-auth-module.js';
+export * from './cookie/cookie-manager.js';
 export * from './decorators.js';
 export * from './errors.js';
 export * from './guard.js';
-export * from './jwt-refresh-token-adapter.js';
+export * from './refresh/jwt-refresh-token-adapter.js';
 export * from './metadata.js';
 export * from './module.js';
-export * from './passport-js.js';
-export * from './refresh-token.js';
+export * from './adapters/passport-js.js';
+export * from './refresh/refresh-token.js';
 export * from './status.js';
 export * from './types.js';
